+++ src/main/java/com/echo/echo/domain/space/SpaceService.java
@@ -1,8 +1,7 @@
 package com.echo.echo.domain.space;
 
-import com.echo.echo.common.exception.CommonCode;
 import com.echo.echo.common.exception.CustomException;
-import com.echo.echo.common.exception.ErrorCode;
+import com.echo.echo.common.exception.codes.ErrorCode;
 import com.echo.echo.domain.space.dto.SpaceRequestDto;
 import com.echo.echo.domain.space.dto.SpaceResponseDto;
 import com.echo.echo.domain.space.entity.Space;
